Refactor Core's Schema trait into its own module
